added fyne test example
