+++ test/filename_test.go
@@ -18,10 +18,28 @@ func TestFunctionName(t *testing.T) {
 	*/
 }
 
-func ExamplTest() { // Test if the following Output is printed
+func ExampleTest() { // Test if the following Output is printed
 	fmt.Println("hello, and") // Print statements actually in the function to be tested
 	fmt.Println("goodbye")
 	// Output:
 	// hello, and
 	// goodbye
 }
+
+/*
+func TestThirdWindowUI(t *testing.T) {
+	out_label, in_entry := makeUI_w3()
+
+	// initial state
+	if out_label.Text != "Hello World!" {
+		t.Error("Incorrect initial value")
+	}
+
+	// user input
+	test.Type(in_entry, "Hi")
+	if out_label.Text != "Hi to you as well!" {
+		t.Error("Incorrect user input handling")
+	}
+}
+
+*/
