--- test/main_test.go
@@ -1,18 +1,0 @@
-package main
-
-/*
-func TestThirdWindowUI(t *testing.T) {
-	out_label, in_entry := makeUI_w3()
-
-	// initial state
-	if out_label.Text != "Hello World!" {
-		t.Error("Incorrect initial value")
-	}
-
-	// user input
-	test.Type(in_entry, "Hi")
-	if out_label.Text != "Hi to you as well!" {
-		t.Error("Incorrect user input handling")
-	}
-}
-*/
